+++ src/commands/utility/team_image.py
